# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NewsBrief Core API
  version: 1.0.0
  description: |
    REST API for authentication, user profile, topics, sources, news ingestion, summarization, translation, and chat.
    Protected endpoints require a Bearer JWT.
servers:
  - url: http://localhost:8080/api/v1
    description: Local
paths:
...
  /admin/create-topics:
    post:
      operationId: createTopic
      tags: [admin]
      summary: Create a topic
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreateTopicRequest' }
      responses:
        '201': { description: Created }
        '403': { description: Forbidden }
...
  /me/topics:
    post:
      operationId: subscribeTopics
      tags: [user]
      summary: Subscribe to multiple topics (batch)
      description: |
        Accepts an array of topic IDs. If the array is empty, the server is a no-op and returns 200.
        If non-empty, the topics are subscribed idempotently and a success message is returned.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscribeTopicsRequest' }
            example:
              topics: [
                "f1d3ff84-12ab-4cde-98ef-abcdef123456",
                "0cc175b9-6a6a-4e2f-bb1d-1234567890ab"
              ]
      responses:
        '200':
          description: Subscribed (or no-op when empty)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
              example: { message: "Subscribed to topic successfully" }
        '401': { description: Unauthorized }
  /me/topics/{topicID}:
    delete:
      operationId: unsubscribeTopic
      tags: [user]
      summary: Unsubscribe from a topic
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: topicID
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Unsubscribed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
              example: { message: "Unsubscribed from topic successfully" }
        '401': { description: Unauthorized }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    MessageResponse:
      type: object
      properties:
        message: { type: string }
...
    SubscribeTopicsRequest:
      type: object
      required: [topics]
      properties:
        topics:
          type: array
          items: { type: string }
...
    AdminCreateTopicRequest:
      type: object
      required: [slug, label]
      properties:
        slug: { type: string, description: Unique topic slug }
        label:
          type: object
          description: Bilingual label for the topic
          required: [en, am]
          properties:
            en: { type: string }
            am: { type: string }